# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hours",
    "description": "Codeborne rust demo",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://cb-rust-demo.herokuapp.com/api",
      "description": "Live"
    },
    {
      "url": "http://localhost:8080/api",
      "description": "Development"
    }
  ],
  "paths": {
    "/hours": {
      "get": {
        "summary": "List all logged hours",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoursList"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Log hours",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewHours"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully logged hours entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hours"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/hours/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "hours entry ID",
          "required": true,
          "schema": {
            "type": "string",
            "format": "UUID"
          }
        }
      ],
      "get": {
        "summary": "Retrieve a single hours entry",
        "responses": {
          "200": {
            "description": "Single hours entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hours"
                }
              }
            }
          },
          "404": {
            "description": "Hours entry not found"
          }
        }
      },
      "delete": {
        "summary": "Remove a single hours entry",
        "responses": {
          "204": {
            "description": "Hours entry was successfully deleted"
          },
          "404": {
            "description": "Hours entry not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "NewHours": {
        "type": "object",
        "required": [
          "employee",
          "date",
          "project",
          "description",
          "hours"
        ],
        "properties": {
          "employee": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "project": {
            "type": "string"
          },
          "storyId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "hours": {
            "type": "integer"
          }
        }
      },
      "Hours": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "UUID"
          },
          "employee": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "project": {
            "type": "string"
          },
          "storyId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "hours": {
            "type": "integer"
          }
        }
      },
      "HoursList": {
        "type": "object",
        "properties": {
          "hours": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Hours"
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}